requirements.txt
google-api-python-client>=2.149.0
google-auth>=2.35.0
google-auth-httplib2>=0.2.0

requests>=2.32.3
beautifulsoup4>=4.12.3

lxml>=5.3.0
python-telegram-bot>=21.0.0,<22

python-dotenv>=1.0.1
tenacity>=9.0.0
rich>=13.7.1
